llm_to_graph/requirements.txt
pydantic==2.11.5
neo4j==5.28.1
python-dotenv==1.1.0

langchain==0.3.25
langchain-google-vertexai==2.0.12
langchain-ollama==0.3.3